24-RestFood
New folders were created (Controller, Exception, Model, Repository). I also added new files inside these folders to have clear structure for the beginning. I did not create the files which we are going to need in next week because of the YAGNI principe, so these files are the files which we are going to work with the next couple of days.