24-Update_MenuController
The Controller part of menu is done and they are no errors and for a stable code I also added exceptions. All the CRUD operations can be used. Service and Controller are full done unless they are going to be some changes from reservation part which could affect the menu parts.